Safari-611.1.10/LayoutTests/fast/text/dangling-surrogates.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<span id="span">
<p id="description"></p>
<div id="console"></div>
<script>
    description("This tests verifies that the test tools can handle a dangling surrogate character");

    var danglingFirst = "\uD803";
    var danglingSecond = "\uDC01";
    
    shouldBe('danglingFirst', '"\uD803"');
    shouldBe('danglingSecond', '"\uDC01"');
</script>
</body>
</html>
