LayoutTests/fast/text/dangling-surrogates-expected.txt

This tests verifies that the test tools can handle a dangling surrogate character

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS danglingFirst is "�"
PASS danglingSecond is "�"
PASS successfullyParsed is true

TEST COMPLETE
